% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcpt-analysis.R
\name{analyze_performance}
\alias{analyze_performance}
\title{Title}
\usage{
analyze_performance(df_pcpt, by_condition = FALSE)
}
\arguments{
\item{df_pcpt}{preprocessed data frame with pebl data}
}
\value{
data.frame
}
\description{
Title
}
