% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcpt-getters.R
\name{get_condition}
\alias{get_condition}
\title{Returns table for only specific trial dirations}
\usage{
get_condition(df_pcpt, trial_duration)
}
\arguments{
\item{trial_duration}{}
}
\value{

}
\description{
Returns table for only specific trial dirations
}
