% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcpt-loading.R
\name{load_experiment}
\alias{load_experiment}
\title{Loads experimental data from a folder. Expects two files _matr.csv
and report _matr.txt.0}
\usage{
load_experiment(folder, preprocess = TRUE)
}
\arguments{
\item{folder}{Folder in which to search for the foiles.}
}
\value{

}
\description{
Loads experimental data from a folder. Expects two files _matr.csv
and report _matr.txt.0
}
