% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcpt-preprocessing.R
\name{preprocess_experiment}
\alias{preprocess_experiment}
\title{Preprocesses some of the columns int he loaded dataframe}
\usage{
preprocess_experiment(df_pcpt)
}
\arguments{
\item{df_pcpt}{data.frame loaded with `\code{\link{load_experiment}}`}
}
\value{
data.frame
}
\description{
Converts column types, adds new columns
}
